feat(pubky): use builder pattern for client::list()

+++ pubky-homeserver/src/routes/public.rs
@@ -88,7 +88,7 @@ pub async fn get(
             public_key,
             path.as_str(),
             params.contains_key("reverse"),
-            params.get("limit").and_then(|l| l.parse::<i32>().ok()),
+            params.get("limit").and_then(|l| l.parse::<u16>().ok()),
             params.get("cursor").map(|cursor| cursor.into()),
         )?;
